+++ utils.py
@@ -1,0 +1,7 @@
+##  Funciones de soporte y utilidad:
+
+import unicodedata
+
+def remove_accents(input_str):
+    nfkd_form = unicodedata.normalize('NFKD', input_str)
+    return "".join([c for c in nfkd_form if not unicodedata.combining(c)])
